Improved tasks

+++ src/main/kotlin/g2901_3000/s2967_minimum_cost_to_make_array_equalindromic/Solution.kt
@@ -19,7 +19,7 @@ class Solution {
             ans1 += abs((previousPalindrome - num))
             ans2 += abs((nextPalindrome - num))
         }
-        return min(ans1.toDouble(), ans2.toDouble()).toLong()
+        return min(ans1, ans2)
     }
 
     private fun getPreviousPalindrome(num: Int): Int {
